// frontend/frontendERP/src/app/api/models/bouteille-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Type } from '../models/type';

/**
 * Creating a bottled product consumes the corresponding raw stock container for the chosen `type`. If raw stock is insufficient an error (400) is returned. Allowed litrages: 1, 0.5, 2, 5 (see /api/commercial/litrages).
 */
export interface BouteilleRequest {
  litrage: number;
  prix: number;
  type: Type;
}