// frontend/frontendERP/src/app/api/models/boite-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { BouteilleRef } from '../models/bouteille-ref';

/**
 * A box (boite) is assembled from bottles according to litrage ratios: - Carton 1l : 15 bouteilles - Carton 0.5l : 30 bouteilles - Carton 2l : 8 bouteilles - Carton 5l : 6 bouteilles The service will validate available bottles and consume them. If insufficient, a 400 error is returned.
 */
export interface BoiteRequest {

  /**
   * Liste des bouteilles; seules les propriétés id sont nécessaires
   */
  bouteilles: Array<BouteilleRef>;
  prix: number;

  /**
   * Peut être 0 (calculé) ou égal au ratio imposé par le litrage
   */
  quantite?: number;
}